+++ src/main/java/a4/streamx_be/services/strategy/RagTTSChatStrategy.java
@@ -1,13 +1,12 @@
-package a4.streamx_be.services.impl;
+package a4.streamx_be.services.strategy;
 
 import a4.streamx_be.configuration.CharacterConfig;
-import a4.streamx_be.domain.dto.ChatType;
+import a4.streamx_be.domain.model.ChatType;
 import a4.streamx_be.domain.dto.request.AIReqDtoV2;
 import a4.streamx_be.domain.dto.response.AIResDtoV3;
-import a4.streamx_be.domain.dto.response.Emotion;
+import a4.streamx_be.domain.model.Emotion;
 import a4.streamx_be.services.ChatStrategy;
-import a4.streamx_be.util.TTSClient;
-import a4.streamx_be.util.impl.FeignTTSService;
+import a4.streamx_be.util.tts.FeignTTSService;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.RequiredArgsConstructor;
