+++ src/main/java/a4/streamx_be/util/tts/FeignTTSClient.java
@@ -1,6 +1,5 @@
-package a4.streamx_be;
+package a4.streamx_be.util.tts;
 
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
