+++ src/main/java/a4/streamx_be/util/tts/TTSClient.java
@@ -1,4 +1,4 @@
-package a4.streamx_be.util;
+package a4.streamx_be.util.tts;
 
 public interface TTSClient {
     String requestAudio(String text);
